% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Performs gene clustering}
\usage{
cluster_genes(se, method = "kmeans", k = 3)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{method}{"kmeans" o "hclust"}

\item{k}{clusters number}
}
\description{
Performs gene clustering
}
