% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Identify outlier genes based on variance Z-score}
\usage{
detect_outliers(se, threshold = 3)
}
\description{
Identify outlier genes based on variance Z-score
}
