% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data and remove low-variance genes}
\usage{
normalize_data(se, method = "log")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{method}{Normalization method ("log" or "quantile")}
}
\description{
Normalize data and remove low-variance genes
}
