% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_expression}
\alias{read_expression}
\title{Read Gene Expression Data}
\usage{
read_expression(path)
}
\arguments{
\item{path}{Path to the csv file.}
}
\value{
A SummarizedExperiment object.
}
\description{
Read Gene Expression Data
}
