% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_cluster_enrichment}
\alias{run_cluster_enrichment}
\title{Gene Set Enrichment Analysis (GSEA) with ID conversion}
\usage{
run_cluster_enrichment(se, cluster_id, org_db = "org.Hs.eg.db")
}
\arguments{
\item{se}{SummarizedExperiment}

\item{cluster_id}{The number of the cluster to be analyzed}

\item{org_db}{The organism's database (default "org.Hs.eg.db" for human)}
}
\description{
Gene Set Enrichment Analysis (GSEA) with ID conversion
}
